pyproject.toml
[project]
name = "exdark"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Wojciech Zarzecki", email = "wojtek.zarzecki2002@gmail.com"},
]
dependencies = [
    "pandas>=2.2.2",
    "opencv-python>=4.10.0.84",
    "albumentations==1.3.1",
    "matplotlib>=3.9.2",
    "torchmetrics[detection]>=1.4.3",
    "numpy==1.26.4",
    "tqdm>=4.66.5",
    "torch>=2.4.1",
    "torchvision>=0.19.1",
    "lightning>=2.4.0",
    "tensorboard>=2.18.0",
    "python-dotenv>=1.0.1",
    "transformers>=4.46.2",
    "wandb>=0.18.6",
    "timm>=1.0.11",
    "hydra-core>=1.3.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true


[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "nbconvert>=7.16.4",
    "pytest>=8.3.4",
    "ruff>=0.8.6",
]
